Exit Idle
Enemy stops being idle and will move continously the direction it is
when it first detects the player. At first glance I thoght this would
make for stupid behaviour (is moving right, enemy is to the left,
continue moving right), but after watching and experimenting, it looks
like it'll be fine.
Lemme know if it's not